+++ __tests__/test-exports.js
@@ -47,6 +47,7 @@ describe('exports', () => {
         "commitHeaderField",
         "commitNthAncestor",
         "commitParent",
+        "createProxyAgent",
         "currentBranch",
         "default",
         "deleteBranch",