+++ src/commands/push.js
@@ -65,6 +65,7 @@ export async function _push({
   delete: _delete = false,
   corsProxy,
   headers = {},
+  agent,
 }) {
   const ref = _ref || (await _currentBranch({ fs, gitdir }))
   if (typeof ref === 'undefined') {
@@ -113,6 +114,7 @@ export async function _push({
     url,
     headers,
     protocolVersion: 1,
+    agent,
   })
   const auth = httpRemote.auth // hack to get new credentials from CredentialManager API
   let fullRemoteRef
@@ -268,6 +270,7 @@ export async function _push({
     auth,
     headers,
     body: [...packstream1, ...packstream2],
+    agent,
   })
   const { packfile, progress } = await GitSideBand.demux(res.body)
   if (onMessage) {